added delete buttons, broke out event listeners into new file
